Add RPath-like support to Windows compiler wrapper (#2)
The Windows dynamic libraries are broken into two parts, an import
library (.lib) and the actual dynamic library (.dll). The import
library is the component utilized when linking a program or library
against the dynamic lib, as it provides the symbol stubs, and tells
the linker where to look for them at runtime via a name in the COFF
format of the import library section associated with the symbol being
resolved. Typically this name is resolved from the qualified path to a
dll at runtime via a complex series of filesystem lookup heuristics.
However, if the name happens to be the absolute path to the
respective DLL, then the linker behaves like it has an RPath and just
uses that absolute path to resolve the DLL.

MSVC's linker does not generate import libraries like this directly,
but includes a tool lib.exe which is capable of doing so: in this PR,
we intercept linker calls, record the libraries that are passed to
the linker, run the linker as normal to generate an initial import
lib, and then use lib.exe to generate a modified import lib with
absolute paths. While lib.exe is capable of computing offsets
correctly for the longer absolute DLL paths, it does not accept path
characters, so it is actually given a mangled name that can later be
mapped to the true path; a final transformation step (largely handled
here in winrpath.cxx) unmangles the paths.

While all linking is done via link.exe, we intercept other compiler
calls (cl/ifx) to allow for insertion of compiler options (e.g.
include directories).

This PR also adds compiler wrapper side handling for relocation
respective to the RPath considerations that need to be made for
successful relocation of binaries.

This PR adds the "relocate" command to the compiler wrapper, which,
when invoked, can either modify a given .lib and its corresponding
dll, or an executable, to ensure it's DLL "rpaths" created by Spack
can be relocated via the binary cache.

On extraction from the binary cache, the "relocate" command will
regenerate import libraries pointing at the new location of their
respective DLLs, and re-write DLLs and EXEs to replace their existing
"rpath" entries to point to the appropriate Spack locations on the
host system onto which they are being relocated.

Spack externals must be "re-written" into Spack specific prefixes
for utilization in this process. Without such an operation, the
binaries will be unable to locate their dependencies. The relocate
command can be utilized to produce an import library of a given
external with it's DLL export references providing an absolute path
to the external DLL being linked to, allowing all Spack builds to
treat externals with the same rpath approach as a native Spack
package. Further, these references to externals can also be
relocated, as above, post relocation from the build cache.