Ensure sections are searched on size boundaries (#5)
Each coff member has a series of sections. Previously our search for
a dll name in those sections was not properly respecting the size of
the section, and was continuing the name search into other sections
and members

This resulted in a bug where the name search found a name in another
member, but because the offset was computed w.r.t the previous members
section, the rename overwrote the next members header, corrupting the
archive.