Env Check: Re-enable validation of Spack build env context (#4)
The shell script version of the Spack compiler wrapper (in the main
Spack repo) validates that there is a well formed Spack env by testing
for the presence of environment variables. Re-enable the same
functionality in the MSVC wrapper. 

Additionally updates the MSVC wrapper validation to reflect the current
state of env validation in the sh wrapper.