LinkerInvocation: refactor argument forwarding approach (#11)
General workflow for generating an "rpath'd" import library:

* Run the linker (link.exe) and allow it to produce a dll and import
  library.
* Parse the linker command line, and extract what we think we need.
  Pass that to the librarian (lib) along with the absolute path to
  the dll as the name.
* Replace the old import lib with the new one.

Previously the wrapper had an ad-hoc approach to forwarding arguments
from the linker to the librarian; this resulted in cases where CLI
args needed by the librarian would be dropped.

This PR dds a map including every argument that is valid for both the
linker and librarian (and makes sense to forward for our case), and
checks that map against the linker CLI to ensure we're forwarding all
arguments verbatim (with handling for quoting).