utils: Strip padding && Special case Spack-built Libraries (#16)
Adds a method to support removing the padding we add to our "RPath"
paths for easier comparison with paths derived from the local
filesystem/environment.

When performing stage -> install prefix relocation, any DLL referenced
by the library we're relocating that is Spack-derived and not part of
the package we're installing already has the appropriate RPath and
therefore does not require relocation. Add logic to special case those
DLLs by checking an environment variable defining the spack install
prefix against the install prefix of the DLLs were considering for
relocation.