Improve escaped quote handling (#18)
Before:

We were re-escaping escaped quotes or quoting strings with reserved
characters, relying on the escaped escapes to handle the reserved
characters, assuming the caller had constructed their strings in
such as way that evaluating the previous escapes would yield a valid
string. This works great for strings like:

"path with a space" -> path with a space 

or 

"\path\" -> "path"

but not so much for

-DPRECROSSESSOR_INCLUDE_PATH="\"path with spaces\"" -> 
"-DPREPROCESSOR_INCLUDE_PATH="path with spaces"".

As you can see this last string gets broken down into
"-DPREPROCESSOR_INCLUDE_PATH=" and "path", "with" "spaces".

Now:

We do both for all strings, ensure more robust passing of complex
strings.

Which results in

-DPRERPROCESSOR_INCLUDE_PATH="\"path with spaces\"" ->
"-DPREPROCESSOR_INCLUDE_PATH=\"path with spaces\"" 

which is then interpreted as -DPREPROCESSOR_INCLUDE_PATH="path with spaces",
which is what we want.

Add a test to ensure we handle this case.

---------

Signed-off-by: John Parent <john.parent@kitware.com>