utils: mangle_name handle long paths
mangle_name previously assumed that the path being passed into pad_path was an appropriate length.
Add handling for long DLL name paths.
Add a check for names longer than 143 characters (DLL name limit in lib/DLL files).
Add handling for names that are longer than 143 characters.
Use Short FIle Names (SFN or 8.3 file names) to attempt to reduce the name length. SFN are a terse, legal naming scheme for Windows paths. However they can be non deterministic, as in cases where there are multiple paths with the same 4 characters, the SFN name will prepend a number. This number's meaning changes as directories/files are added and removed.
However, relative to the install/stage prefix of a package, no new directories are being added at link/relocate time.
The shortening strategy is then to split the spack install/stage prefix from the relative path to the DLL/Lib, and only compute the SFN for the invariant (relative) portion of the path. Then rejoin.

If this partially shortened path is still to long, we stop attempting to do anything and request the user move to a shorter prefix.

Signed-off-by: John Parent <john.parent@kitware.com>