Support exes with exports (#22)
Previously we were only handling the case where DLLs were responsible
for exporting symbols, but now we handle the case where Exes may do
so as well. We accomplish this by genericizing all checks for shared
libraries to check for DLLs and EXEs and refer to them as PE.

Add a test for this.

---------

Signed-off-by: John Parent <john.parent@kitware.com>