Improve def support (#21)
Previously we were ignore def files specified to the linker. This can
cause binaries to be built improperly as this is a mechanism some
projects use to define the symbols they're interested in exporting,
in addition to a specification of exe and dll names.

The EXPORTS keyword in a def file denotes symbols to be exported from
a DLL and therefor to be included in an import library. The NAME and
LIBRARY keywords denote executable and dll names respectively (but do
not change the import library names).

This PR:

* Forwards def files properly from the linker invocation to our
  second pass of the librarian tool
* Filters the def file we pass to the librarian tool to remove the
  NAME and LIBRARY keywords. Without this, the keywords specified in
  the def file override our mangled name on the command line, and the
  mangling (which we rely on to inject the absolute path) does not
  work from inside the def file. All EXPORTS and every other line of
  the def file is copied verbatim.
* Assigns the name of the produced dll/exe and import library
  appropriately based on the command line and the contents of the def
  file's LIBRARY and NAME keywords.
* Adds a test that specifies a DLL name and some symbols to export in
  a def file, compiles a library with the def file, and then proceeds
  to link and run and executable against our def based library to
  validate the above behavior.

Signed-off-by: John Parent <john.parent@kitware.com>